hooked up redux to header
