+++ src/containers/HeaderContainer
@@ -1,0 +1,21 @@
+import Header from '../components/Header';
+import {login} from '../actions/loginActions';
+import {connect} from 'react-redux';
+
+
+const mapStateToProps = (state) => {
+    return {
+        login: state.login,
+        logout:state.logout
+    };
+};
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        doLogin: (userName, password) => {
+            dispatch(login(userName,password));
+        }
+    };
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(Header);
